Studio: move an untouched Whisper Small setting to the recommended model
Changing DEFAULT_STT_MODEL only reached fresh installs. Anyone who had opened
Voice settings before had "small" saved, so the store kept it and the download
prompt still offered Whisper Small on a machine with no model at all.

A v1 migration moves a saved "small", the old default, to the recommended
model. Any other saved model was picked on purpose and is left alone, and
choosing Small again persists at v1 so it sticks.

The model list and the migration move to stt-model-catalog, which imports
nothing from the app, so the node test runner can exercise them directly. The
store re-exports it, so no consumer changes.
